requirement_local.txt
langchain==0.3.27
wikipedia==1.4.0
langchain-community==0.3.29
langchain-openai==0.3.31
langchain-anthropic==0.3.19
python-dotenv==1.1.1
pydantic==2.11.7
langchain-ollama==0.3.7
langchain-chroma==0.2.5
pandas==2.3.2